requirements.txt
streamlit
numpy
pandas
matplotlib
scipy
scikit-learn
python-docx

